pc_app/requirements.txt
pyserial>=3.5
dearpygui>=2.0.0
numpy>=1.24.0
